Move time parsing to avoid future import loops
