+++ go/vt/vtgate/evalengine/api_literal.go
@@ -24,10 +24,10 @@ import (
 	"unicode/utf8"
 
 	"vitess.io/vitess/go/mysql/collations"
+	"vitess.io/vitess/go/mysql/datetime"
 	"vitess.io/vitess/go/mysql/decimal"
 	"vitess.io/vitess/go/sqltypes"
 	querypb "vitess.io/vitess/go/vt/proto/query"
-	"vitess.io/vitess/go/vt/sqlparser"
 )
 
 // NullExpr is just what you are lead to believe
@@ -103,7 +103,7 @@ func NewLiteralString(val []byte, collation collations.TypedCollation) *Literal 
 
 // NewLiteralDateFromBytes returns a literal expression.
 func NewLiteralDateFromBytes(val []byte) (*Literal, error) {
-	_, err := sqlparser.ParseDate(string(val))
+	_, err := datetime.ParseDate(string(val))
 	if err != nil {
 		return nil, err
 	}
@@ -113,7 +113,7 @@ func NewLiteralDateFromBytes(val []byte) (*Literal, error) {
 // NewLiteralTimeFromBytes returns a literal expression.
 // it validates the time by parsing it and checking the error.
 func NewLiteralTimeFromBytes(val []byte) (*Literal, error) {
-	_, err := sqlparser.ParseTime(string(val))
+	_, err := datetime.ParseTime(string(val))
 	if err != nil {
 		return nil, err
 	}
@@ -123,7 +123,7 @@ func NewLiteralTimeFromBytes(val []byte) (*Literal, error) {
 // NewLiteralDatetimeFromBytes returns a literal expression.
 // it validates the datetime by parsing it and checking the error.
 func NewLiteralDatetimeFromBytes(val []byte) (*Literal, error) {
-	_, err := sqlparser.ParseDateTime(string(val))
+	_, err := datetime.ParseDateTime(string(val))
 	if err != nil {
 		return nil, err
 	}
